Move our own JavaScript into Webpacker
Here we move the custom JS code we have in the application from
`app/assets/javascripts` to `app/javascript/packs`, so that it gets
compiled by Webpack rather than by the legacy Rails asset pipeline.

JavaScript source code lives in `app/javascript/src`. Files in
`app/javascript/packs` define bundles of JS that are actually loaded
into our pages; these `packs` files should import files from `src`.

Files in `packs` can be loaded into pages using the
`javascript_pack_tag` view helper.

Third-party JS is still loaded using `//= require` comments in the
original application file and are processed by the asset pipeline.
