+++ app/javascript/packs/application.js
@@ -15,4 +15,5 @@
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 
-console.log('Hello World from Webpacker')
+import cookieConsent from "../src/cookie-consent"
+import toggleProvidingOrgFields from "../src/toggle-providing-org-fields"
