+++ app/javascript/src/cookie-consent.js
